# contracts/web-api.openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "Cherry OJ Web API",
    "version": "0.2.0",
    "description": "浏览器与 Gateway 之间的公开 REST 契约。请求使用 endpoint DTO，普通 JSON 成功响应使用 ApiSuccess，失败响应使用 RFC 9457 Problem Details。 管理部署接口的 503 可返回固定节点错误：NO_ONLINE_JUDGE_NODE, JUDGE_NODE_UNREACHABLE, JUDGE_NODE_DATA_REJECTED, JUDGE_NODE_RECEIPT_MISMATCH。详情由固定文案生成，不透传节点正文；readiness checks 的 ONLINE_JUDGE_NODE 表示 ACTIVE 环境存在在线节点，DEPLOYMENT 仅在该在线节点持有匹配摘要回执时通过。"
  },
...
  "paths": {
...
    "/api/submissions": {
      "get": {
        "operationId": "listProblemSubmissions",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "X-Expected-User-Id",
            "in": "header",
            "required": true,
            "description": "编辑器所属账号的一致性前置条件；Gateway 与已验证会话比较，此值不参与授权。",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "problemId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2147483647,
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "verdict",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "AC",
                "WA",
                "PE",
                "CE",
                "RE",
                "TLE",
                "MLE",
                "OLE",
                "SE"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "仅当前账号的提交；恢复操作不创建新记录。",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionListSuccess"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      },
      "post": {
        "operationId": "createSubmission",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "X-Expected-User-Id",
            "in": "header",
            "required": true,
            "description": "编辑器所属账号的一致性前置条件；Gateway 与已验证会话比较，此值不参与授权。",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "仅当前账号的提交；恢复操作不创建新记录。",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionSuccess"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          },
          "201": {
            "description": "首次受理；同键同内容返回 200，同键不同内容或版本过期返回 409。",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionSuccess"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubmissionRequest"
              }
            }
          }
        }
      }
    },
    "/api/submissions/{id}/source": {
      "get": {
        "operationId": "getSubmissionSource",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Expected-User-Id",
            "in": "header",
            "required": true,
            "description": "编辑器所属账号的一致性前置条件；Gateway 与已验证会话比较，此值不参与授权。",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "仅当前账号的提交；恢复操作不创建新记录。",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionSourceSuccess"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      }
    },
    "/api/submissions/{id}": {
      "get": {
        "operationId": "getSubmission",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "仅当前账号的提交；恢复操作不创建新记录。",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionSuccess"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      }
    },
    "/api/submission-requests/{key}": {
      "get": {
        "operationId": "getSubmissionRequest",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "仅当前账号的提交；恢复操作不创建新记录。",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionSuccess"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "CHERRY_SESSION",
        "description": "浏览器随机 Session Cookie；不是内部 JWT。"
      }
    },
    "parameters": {
      "CsrfToken": {
        "name": "X-CSRF-Token",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 16,
          "maxLength": 512
        }
      },
...
    },
    "headers": {
      "XRequestId": {
        "description": "Gateway 生成的公开请求标识；必须与响应 body 的 meta.requestId 相同。",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/RequestId"
        }
      }
    },
    "responses": {
...
      "ApiProblemResponse": {
        "description": "符合 RFC 9457 且经过 Gateway 脱敏的错误响应。实际 HTTP status 与 body.status 相同。",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/XRequestId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ApiProblem"
            },
            "example": {
              "type": "urn:cherry-oj:problem:validation-failed",
              "title": "请求参数校验失败",
              "status": 422,
              "detail": "请检查标记字段。",
              "instance": "urn:cherry-oj:request:req_01K37XZ3MFXBK92WMG67G4XFN0",
              "code": "VALIDATION_FAILED",
              "meta": {
                "requestId": "req_01K37XZ3MFXBK92WMG67G4XFN0"
              },
              "violations": [
                {
                  "path": "title",
                  "code": "SIZE",
                  "message": "标题长度必须在 1 到 120 之间。"
                }
              ]
            }
          }
        }
      }
    },
    "schemas": {
      "SubmissionSourceData": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "submissionId",
          "problemId",
          "languageId",
          "source"
        ],
        "properties": {
          "submissionId": {
            "type": "string",
            "format": "uuid"
          },
          "problemId": {
            "type": "string",
            "format": "uuid"
          },
          "languageId": {
            "type": "string",
            "const": "cpp"
          },
          "source": {
            "type": "string",
            "maxLength": 262144,
            "description": "本人提交时的原始代码，UTF-8 不超过 256 KiB；不含内部送判数据。"
          }
        }
      },
      "SubmissionSourceSuccess": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SubmissionSourceData"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "SubmissionListSuccess": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/SubmissionData"
            }
          },
          "meta": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ApiMeta"
              },
              {
                "type": "object",
                "required": [
                  "pagination"
                ],
                "properties": {
                  "pagination": {
                    "$ref": "#/components/schemas/PagePagination"
                  }
                }
              }
            ]
          }
        }
      },
      "CreateSubmissionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "problemId",
          "languageId",
          "source"
        ],
        "properties": {
          "problemId": {
            "type": "string",
            "format": "uuid"
          },
          "languageId": {
            "type": "string",
            "const": "cpp"
          },
          "source": {
            "type": "string",
            "minLength": 1,
            "maxLength": 262144,
            "x-max-utf8-bytes": 262144,
            "description": "完整 ACM 程序；服务端另校验 UTF-8 字节上限。"
          }
        }
      },
      "SubmissionData": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "problemId",
          "problemTitle",
          "languageId",
          "status",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "problemId": {
            "type": "string",
            "format": "uuid"
          },
          "problemTitle": {
            "type": "string"
          },
          "languageId": {
            "type": "string",
            "const": "cpp"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "JUDGING",
              "DONE"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "verdict": {
            "type": "string",
            "enum": [
              "AC",
              "WA",
              "PE",
              "TLE",
              "MLE",
              "OLE",
              "RE",
              "CE",
              "SE"
            ]
          },
          "cpuNs": {
            "type": "integer",
            "minimum": 0
          },
          "memoryBytes": {
            "type": "integer",
            "minimum": 0
          },
          "passedCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000
          },
          "executedCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000
          },
          "totalCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000
          },
          "message": {
            "type": "string",
            "maxLength": 8192
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "SubmissionSuccess": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SubmissionData"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "RequestId": {
        "type": "string",
        "pattern": "^req_[A-Za-z0-9_-]{16,64}$",
        "description": "Gateway 生成的 opaque public request ID，不等同于 session、trace 或幂等键。"
      },
      "ApiMeta": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "requestId"
        ],
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/RequestId"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "examples": [
          {
            "requestId": "req_01K37XZ3MFXBK92WMG67G4XFN0"
          }
        ]
      },
      "Pagination": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/CursorPagination"
          },
          {
            "$ref": "#/components/schemas/PagePagination"
          }
        ],
        "discriminator": {
          "propertyName": "kind",
          "mapping": {
            "cursor": "#/components/schemas/CursorPagination",
            "page": "#/components/schemas/PagePagination"
          }
        }
      },
      "CursorPagination": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "kind",
          "nextCursor",
          "hasMore"
        ],
        "properties": {
          "kind": {
            "const": "cursor",
            "type": "string"
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ],
            "minLength": 1,
            "maxLength": 2048,
            "description": "不透明游标；客户端不得解析。"
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "examples": [
          {
            "kind": "cursor",
            "nextCursor": null,
            "hasMore": false
          }
        ]
      },
      "PagePagination": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "kind",
          "page",
          "size",
          "totalElements",
          "totalPages"
        ],
        "properties": {
          "kind": {
            "const": "page",
            "type": "string"
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "size": {
            "type": "integer",
            "minimum": 1
          },
          "totalElements": {
            "type": "integer",
            "minimum": 0
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0
          }
        },
        "examples": [
          {
            "kind": "page",
            "page": 1,
            "size": 20,
            "totalElements": 0,
            "totalPages": 0
          }
        ]
      },
      "FieldViolation": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "path",
          "code",
          "message"
        ],
        "properties": {
          "path": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512
          },
          "code": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]{0,63}$"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1024
          }
        },
        "examples": [
          {
            "path": "title",
            "code": "SIZE",
            "message": "标题长度必须在 1 到 120 之间。"
          }
        ]
      },
      "ApiProblem": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "type",
          "title",
          "status",
          "code",
          "meta"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2048
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          },
          "code": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]{0,63}$"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          },
          "violations": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/FieldViolation"
            }
          }
        },
        "examples": [
          {
            "type": "urn:cherry-oj:problem:validation-failed",
            "title": "请求参数校验失败",
            "status": 422,
            "detail": "请检查标记字段。",
            "instance": "urn:cherry-oj:request:req_01K37XZ3MFXBK92WMG67G4XFN0",
            "code": "VALIDATION_FAILED",
            "meta": {
              "requestId": "req_01K37XZ3MFXBK92WMG67G4XFN0"
            },
            "violations": [
              {
                "path": "title",
                "code": "SIZE",
                "message": "标题长度必须在 1 到 120 之间。"
              }
            ]
          }
        ]
      },
...
    }
  }
}